+++ src/main/java/one/util/streamex/IntStreamEx.java
@@ -429,7 +429,7 @@ public class IntStreamEx extends BaseStreamEx<Integer, IntStream, Spliterator.Of
      * @since 0.0.8
      */
     public IntStreamEx reverseSorted() {
-        IntUnaryOperator inv = x -> -1-x;
+        IntUnaryOperator inv = x -> ~x;
         return new IntStreamEx(stream().map(inv).sorted().map(inv), context);
     }