+++ CHANGES.md
@@ -1,7 +1,7 @@
 # StreamEx changes
 
 ### 0.5.5
-* [#63] Optimized IntStreamEx/LongStreamEx.reverseSorted(): much faster now and take less memory.
+* [#63] Optimized `IntStreamEx/LongStreamEx/DoubleStreamEx.reverseSorted()`: much faster now and take less memory.
 
 ### 0.5.4
 * [#10] Added: `chain()` method to all stream types allowing fluently chaining custom operations.